Refactor featurizers
