+++ msmbuilder/featurizer/__init__.py
@@ -1,5 +1,7 @@
 from __future__ import absolute_import
 
+from .base import Featurizer, featurize_all
 from .feature_union import FeatureUnion
 from .featurizer import *
 from .indices import get_atompair_indices
+from .rmsd import RMSDFeaturizer
